refactor: move central source files into a specialized namespace
- Move the core source files to the special namespace, keeping the logic untouched
- Changed the variable names in Hilbert wrapper, for easier extending the features
