+++ .vs/Elliptic Curve Primality Proving/v18/DocumentLayout.json
@@ -1,0 +1,82 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{96330448-19DB-4F3B-A205-723D66481901}|Elliptic Curve Primality Proving\\Elliptic Curve Primality Proving.csproj|C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\elliptic curve primality proving\\mainwindow.xaml.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{96330448-19DB-4F3B-A205-723D66481901}|Elliptic Curve Primality Proving\\Elliptic Curve Primality Proving.csproj|solutionrelative:elliptic curve primality proving\\mainwindow.xaml.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{96330448-19DB-4F3B-A205-723D66481901}|Elliptic Curve Primality Proving\\Elliptic Curve Primality Proving.csproj|c:\\users\\eduar\\documents\\github\\ellipticcurveprimalityproving\\elliptic curve primality proving\\core\\atkin.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{96330448-19DB-4F3B-A205-723D66481901}|Elliptic Curve Primality Proving\\Elliptic Curve Primality Proving.csproj|solutionrelative:elliptic curve primality proving\\core\\atkin.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{96330448-19DB-4F3B-A205-723D66481901}|Elliptic Curve Primality Proving\\Elliptic Curve Primality Proving.csproj|C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\elliptic curve primality proving\\mainwindow.xaml||{F11ACC28-31D1-4C80-A34B-F4E09D3D753C}",
+      "RelativeMoniker": "D:0:0:{96330448-19DB-4F3B-A205-723D66481901}|Elliptic Curve Primality Proving\\Elliptic Curve Primality Proving.csproj|solutionrelative:elliptic curve primality proving\\mainwindow.xaml||{F11ACC28-31D1-4C80-A34B-F4E09D3D753C}"
+    }
+  ],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": [
+        {
+          "DockedWidth": 200,
+          "SelectedChildIndex": 4,
+          "Children": [
+            {
+              "$type": "Bookmark",
+              "Name": "ST:0:0:{1c4feeaa-4718-4aa9-859d-94ce25d182ba}"
+            },
+            {
+              "$type": "Bookmark",
+              "Name": "ST:128:0:{116d2292-e37d-41cd-a077-ebacac4c8cc4}"
+            },
+            {
+              "$type": "Bookmark",
+              "Name": "ST:128:0:{75188d03-9892-4ae2-abf1-207126247ce5}"
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 1,
+              "Title": "Atkin.cs",
+              "DocumentMoniker": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\Elliptic Curve Primality Proving\\Core\\Atkin.cs",
+              "RelativeDocumentMoniker": "Elliptic Curve Primality Proving\\Core\\Atkin.cs",
+              "ToolTip": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\Elliptic Curve Primality Proving\\Core\\Atkin.cs",
+              "RelativeToolTip": "Elliptic Curve Primality Proving\\Core\\Atkin.cs",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAgAAAAsAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2026-01-18T12:54:55.94Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "MainWindow.xaml.cs",
+              "DocumentMoniker": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\Elliptic Curve Primality Proving\\MainWindow.xaml.cs",
+              "RelativeDocumentMoniker": "Elliptic Curve Primality Proving\\MainWindow.xaml.cs",
+              "ToolTip": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\Elliptic Curve Primality Proving\\MainWindow.xaml.cs",
+              "RelativeToolTip": "Elliptic Curve Primality Proving\\MainWindow.xaml.cs",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAUAAAAAAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-10-30T18:56:49.426Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 2,
+              "Title": "MainWindow.xaml",
+              "DocumentMoniker": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\Elliptic Curve Primality Proving\\MainWindow.xaml",
+              "RelativeDocumentMoniker": "Elliptic Curve Primality Proving\\MainWindow.xaml",
+              "ToolTip": "C:\\Users\\eduar\\Documents\\GitHub\\EllipticCurvePrimalityProving\\Elliptic Curve Primality Proving\\MainWindow.xaml",
+              "RelativeToolTip": "Elliptic Curve Primality Proving\\MainWindow.xaml",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.003549|",
+              "WhenOpened": "2025-10-30T18:56:49.524Z",
+              "EditorCaption": ""
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
